addr fb: move functions in partials to custom-domain-app-runner
